requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
qrcode==7.4.2
Pillow==10.2.0
python-dotenv==1.0.1
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
gunicorn==21.2.0
